requirements.txt
Django==4.0.4
djangorestframework==3.13.1
wikipedia==1.4.0
